requirements.txt
pyyaml
requests
requests-mock
pytest